# [inductor] set correct precompile start time