@henrylhtsang 2025-04-28
One (non-important) follow-up might be to figure out why some triton kernel takes so long to compile. Maybe has something to do with cache hit / not hit?